--- a/lib/netstd/Benchmarks/Thrift.Benchmarks/Program.cs
+++ b/lib/netstd/Benchmarks/Thrift.Benchmarks/Program.cs
@@ -1,0 +1,12 @@
+using BenchmarkDotNet.Running;
+
+namespace Thrift.Benchmarks
+{
+	internal static class Program
+	{
+		public static void Main(string[] args)
+		{
+			BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
+		}
+	}
+}