Uses ArrayPool to save on buffer allocations when reading/writing strings.

Related to #2043.