@CodeBlanch 2020-03-12
I made a couple changes that might be unpopular but worth doing IMO:

1) Added benchmark project. This was to isolate the changes in order to more accurately measure the performance. Can be built upon in the future to measure other things.
2) Multi-target .NET Standard 2.1 on the Thrift library to access the ReadAsync ValueTask Stream API additions. Greatly reduces allocations performing reads.

### .NET 4.8 Benchmarks
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4150.0), X64 RyuJIT
  DefaultJob : .NET Framework 4.8 (4.8.4150.0), X64 RyuJIT


```
|      Method | NumberOfOperationsPerIteration |     Mean |     Error |    StdDev |    Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |------------------------------- |---------:|----------:|----------:|---------:|------:|------:|----------:|
| WriteString |                          10000 | 2.920 ms | 0.0581 ms | 0.0714 ms |  74.2188 |     - |     - | 470.15 KB |
|  ReadString |                          10000 | 5.939 ms | 0.1166 ms | 0.1780 ms | 367.1875 |     - |     - | 2272.4 KB |

After:
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4150.0), X64 RyuJIT
  DefaultJob : .NET Framework 4.8 (4.8.4150.0), X64 RyuJIT


```
|      Method | NumberOfOperationsPerIteration |     Mean |     Error |    StdDev |    Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |------------------------------- |---------:|----------:|----------:|---------:|------:|------:|----------:|
| WriteString |                          10000 | 3.322 ms | 0.0628 ms | 0.0645 ms |        - |     - |     - |         - |
|  ReadString |                          10000 | 5.863 ms | 0.0941 ms | 0.0834 ms | 367.1875 |     - |     - | 2326867 B |

### .NET Core 3.1 Benchmarks
Before:
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT
  DefaultJob : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT


```
|      Method | NumberOfOperationsPerIteration |     Mean |     Error |    StdDev |    Gen 0 | Gen 1 | Gen 2 |  Allocated |
|------------ |------------------------------- |---------:|----------:|----------:|---------:|------:|------:|-----------:|
| WriteString |                          10000 | 1.359 ms | 0.0222 ms | 0.0197 ms |  56.6406 |     - |     - |  468.75 KB |
|  ReadString |                          10000 | 3.322 ms | 0.0376 ms | 0.0352 ms | 257.8125 |     - |     - | 2109.42 KB |

After:
``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9880H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.101
  [Host]     : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT
  DefaultJob : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT


```
|      Method | NumberOfOperationsPerIteration |     Mean |     Error |    StdDev |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |------------------------------- |---------:|----------:|----------:|--------:|------:|------:|----------:|
| WriteString |                          10000 | 1.614 ms | 0.0315 ms | 0.0420 ms |       - |     - |     - |         - |
|  ReadString |                          10000 | 3.812 ms | 0.0092 ms | 0.0082 ms | 85.9375 |     - |     - |  720000 B |