@Jens-G 2020-03-13
https://devblogs.microsoft.com/dotnet/announcing-net-standard-2-1/

> Given many of the API additions in .NET Standard 2.1 require runtime changes in order to be meaningful, .NET Framework 4.8 will remain on .NET Standard 2.0 rather than implement .NET Standard 2.1.

IIRC this will break then compatibility with NET Framework. That's why we reset it to 2.0.

https://issues.apache.org/jira/browse/THRIFT-4990