@CodeBlanch 2020-03-13
> IIRC this will break then compatibility with NET Framework. That's why we reset it to 2.0.

Your recall is 100% spot-on, that's why I didn't _change_ the target, I _added_ one. Subtle, but important difference. The library is now targetting .NET Standard 2.0 & .NET Standard 2.1 at the same time. When you compile you get 2 DLLs, one for each. Both can be included in the NuGet and then consumers can use whichever they support.

You can see this right now in practice in the Benchmark project. It is a Console app (EXE) multi-targeting .NET Framework 4.8 and .NET Core 3.1. When compiled, two exes. One will use the .NET Standard 2.0 path and the other .NET Standard 2.1.

The idea being, we support people on older versions but if you are using .NET Core 3+, we can take advantage of the performance improvements available.